Primer Examen Parcial
primer examen

+++ src/test/java/com/example/PrimerExamen/service/iCantanteService.java
@@ -8,7 +8,7 @@ import java.util.List;
  */
 public interface iCantanteService {
     public List<cantante> getALLCantante();
-    public cantante getCantanteBylId(int idCantante);
+    public cantante getCantanteBylId(long idCantante);
     public void saveCantante(cantante cantante);
     public void delete(long idCantante);
 }
